backend/requirements.txt
boto3
python-dotenv
flask
flask-cors


